+++ app/models/sbmt/outbox/item.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Sbmt
-  module Outbox
-    Item = OutboxItem
-  end
-end