+++ lib/sbmt/outbox/engine.rb
@@ -33,21 +33,6 @@ module Sbmt
           c.batch_process_middlewares.push("Sbmt::Outbox::Middleware::Sentry::TracingBatchProcessMiddleware")
           c.item_process_middlewares.push("Sbmt::Outbox::Middleware::Sentry::TracingItemProcessMiddleware")
         end
-
-        def c.redis_servers=(val)
-          val = val.first
-
-          if val.is_a?(String)
-            self.redis = {url: val}
-          elsif val.respond_to?(:_client)
-            conf = val._client.config
-            self.redis = {url: conf.server_url, username: conf.username, password: conf.password}
-          else
-            raise ArgumentError, "Outbox `redis_servers=` config option is deprecated. Please use `redis=` with a Hash"
-          end
-
-          warn "🔥 Outbox `redis_servers=` config option is deprecated. Please use `redis=` with a Hash. Called from #{caller(1..1).first}"
-        end
       end
 
       rake_tasks do